pull redis out of app.ts for testing
