+++ tests/routes/index.test.ts
@@ -2,6 +2,7 @@ import app from "../../src/app";
 import request from "supertest";
 import mongoose from "mongoose";
 import {logger} from "../../src/logging";
+import redisClient from "../../src/redis"
 
 describe("GET /", () => {
     afterAll((done) => {
@@ -9,6 +10,8 @@ describe("GET /", () => {
             logger.info("Closing the DB connection...")
             done()
         })
+        redisClient.quit()
+
     })
     it("to be json response with the correct message.",  async () => {
        const res = await request(app)
